Moved connect_server_stdin to multiprocessed backend since it isn't used anywhere else
